refactor(racing-car): InputValidation 클래스 통합
