+++ src/main/java/racingcar/InputValidation.java
@@ -3,11 +3,12 @@ package racingcar;
 import java.util.HashSet;
 import java.util.Set;
 
-public class CarInputValidation {
+public class InputValidation {
     private static final int MIN_CAR_NAME_LENGTH = 0;
     private static final int MAX_CAR_NAME_LENGTH = 5;
     private static final String ERROR_CAR_NAME = "[ERROR] 경주할 자동차 이름이 올바르게 입력되지 않았습니다.";
     private static final String ERROR_CAR_DUPLICATE = "[ERROR] 경주할 자동차 이름이 중복됩니다.";
+    private static final String ERROR_TRIAL_NUMBER = "[ERROR] 시도 횟수는 자연수여야 합니다.";
 
     public static String[] checkCarNameInput(String userInput) {
         String[] cars = userInput.split(",");
@@ -49,6 +50,34 @@ public class CarInputValidation {
             }
         } catch (UserInputException e) {
             System.err.println(ERROR_CAR_NAME);
+            System.exit(0);
+        }
+    }
+
+    public static int checkTrialInput(String userInput) {
+        int trialInput = checkIntegerInput(userInput);
+        checkNaturalNumber(trialInput);
+        return trialInput;
+    }
+
+    private static int checkIntegerInput(String userInput) {
+        int trialInput = 0;
+        try {
+            trialInput = Integer.parseInt(userInput);
+        } catch (NumberFormatException e) {
+            System.err.println(ERROR_TRIAL_NUMBER);
+            System.exit(0);
+        }
+        return trialInput;
+    }
+
+    private static void checkNaturalNumber(int trialInput) {
+        try {
+            if(trialInput <= 0) {
+                throw new UserInputException();
+            }
+        } catch (UserInputException e) {
+            System.err.println(ERROR_TRIAL_NUMBER);
             System.exit(0);
         }
     }
